Refactor icons story to use just HTML.
